Committing renames and refactors of Project 0 Files
Refactored Project 0 code so that the BankMember class and the BankAccount class would each have
their own DAO, Oracle, and Service classes. Performed further testing on the Application class
which increased the number of entries in the accompanying log file
